backend/requirements.txt
asgiref==3.2.5
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0.4
django-filter==2.2.0
django-rest-framework==0.1.0
djangochannelsrestframework==0.0.3
djangorestframework==3.11.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
Markdown==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==19.10.0
txaio==20.1.1
zope.interface==4.7.2
